requirements.txt
pandas~=1.5.2
folium~=0.14.0
branca~=0.6.0
Jinja2~=3.1.2
geopandas~=0.12.2
numpy~=1.24.0
mysql-connector-python~=8.0.31
matplotlib~=3.6.2
seaborn~=0.12.1
adjustText~=0.7.3
geopy~=2.3.0